# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ErrorResponse:
    description: Error response structure
    properties:
      code:
        example: INVALID_INPUT
        type: string
      error:
        example: bad_request
        type: string
      message:
        example: Invalid request parameters
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: kk
    url: https://opensource.org/licenses/MIT
  description: |-
    kkArtifact is a modern artifact management and synchronization system designed to replace rsync + SSH deployment workflows.
    The system provides multi-project and multi-app artifact storage with hash-based versioning, HTTP-based API for artifact upload, download, and management.
    Token-based authentication with fine-grained permissions (Global/Project/App scopes).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: kkArtifact API
  version: "1.0"
paths:
...
  /config:
    get:
      consumes:
      - application/json
      description: Get the global configuration (e.g., version retention limit)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get config
      tags:
      - config
    put:
      consumes:
      - application/json
      description: Update the global configuration
      parameters:
      - description: Configuration updates
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Update config
      tags:
      - config
...
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"